Modify loophole--test-with-keyboard-events to support batch-mode
Reading minibuffer in batch-mode requires standard input, and
unread-command-events are ignored.
loophole--test-with-keyboard-events now advise read-from-minibuffer as
returning a string that is a portion of specified keyboard events
delimited by newline.
Consequently, functions to read minibuffer is made to be independent
of standard input and works with the keyboard events specified to
loophole--test-with-keyboard-events.